requirements.txt

#pip install -r requirements.txt
python-dotenv
google-generativeai
psutil
torch
transformers
pydantic
fastapi
uvicorn
